Migrate users management commands to the auth app
